+++ lib/registries/addon/drafts/draft-shim/route.ts
@@ -3,7 +3,8 @@ import Route from '@ember/routing/route';
 import { DefaultPage } from 'ember-osf-web/utils/page-param';
 
 export default class DraftIndexRoute extends Route {
-    model(this: DraftIndexRoute, params: { id: string }) {
+    beforeModel() {
+        const params: { id?: string } = this.paramsFor(this.routeName);
         return this.replaceWith('drafts.draft', params.id, DefaultPage);
     }
 }