+++ mirage/config.ts
@@ -104,11 +104,6 @@ export default function(this: Server) {
     osfNestedResource(this, 'registration', 'forks', { except: ['create'] });
     this.post('/registrations/:id/forks', forkRegistration);
 
-    osfResource(this, 'draft-registration', {
-        only: ['index', 'show'],
-        path: '/draft_registrations',
-    });
-
     osfNestedResource(this, 'registration', 'contributors', { defaultSortKey: 'index' });
     osfNestedResource(this, 'registration', 'bibliographicContributors', {
         only: ['index'],